Adding example of dora not getting the latest data when a queue_size is set

+++ tests/queue_size_latest_data_python/send_data.py
@@ -1,10 +1,11 @@
 from dora import Node
 import time
 import pyarrow as pa
+import numpy as np
 
 node = Node()
 
 for i in range(10):
     now = time.time()
-    node.send_output("data", pa.array([now]))
+    node.send_output("data", pa.array([np.uint64(now)]))
     time.sleep(1)
